+++ src/backend/resolvers/mutations/organization/mbm/lbm/relational-resolvers/index.ts
@@ -1,13 +1,13 @@
-// import serviceAndServiceCategoryResolvers from './serviceAndServiceCategory'
-// import lbmAndServiceResolvers from './lbmAndService'
+import serviceAndServiceCategoryResolvers from './serviceAndServiceCategory'
+import lbmAndServiceResolvers from './lbmAndService'
 // import lbmAndPersonResolvers from './lbmAndPerson'
 import lbmAndLbmTypeResolvers from './lbmAndLbmType'
 // import lbmAndKeyEventResolvers from './lbmAndKeyEvent'
 import lbmAndPayer from './lbmAndPayer'
 
 export default {
-  // ...serviceAndServiceCategoryResolvers,
-  // ...lbmAndServiceResolvers,
+  ...serviceAndServiceCategoryResolvers,
+  ...lbmAndServiceResolvers,
   // ...lbmAndPersonResolvers,
   ...lbmAndLbmTypeResolvers,
   // ...lbmAndKeyEventResolvers,