+++ src/backend/resolvers/mutations/organization/mbm/lbm/relational-resolvers/lbmAndService/index.ts
@@ -1,0 +1,5 @@
+import connectLbmAndLbmService from './replace'
+
+export default {
+  connectLbmAndLbmService,
+}