+++ src/backend/resolvers/mutations/organization/mbm/lbm/relational-resolvers/serviceAndServiceCategory/index.ts
@@ -1,0 +1,5 @@
+import connectLbmServiceAndLbmServiceCategory from './upsert'
+
+export default {
+  connectLbmServiceAndLbmServiceCategory,
+}