+++ src/backend/typeDefs/mutations/organization/mbm/lbm/relational-typeDefs/index.ts
@@ -1,13 +1,13 @@
-// import serviceAndServiceCategoryTypeDefs from './serviceAndServiceCategory'
-// import lbmAndServiceTypeDefs from './lbmAndService'
+import serviceAndServiceCategoryTypeDefs from './serviceAndServiceCategory'
+import lbmAndServiceTypeDefs from './lbmAndService'
 import lbmAndLbmTypeTypeDefs from './lbmAndLbmType'
 // import lbmAndPersonTypeDefs from './lbmAndPerson'
 // import lbmAndKeyEvent from './lbmAndKeyEvent'
 import lbmAndPayer from './lbmAndPayer'
 
 export default [
-  // ...serviceAndServiceCategoryTypeDefs,
-  // ...lbmAndServiceTypeDefs,
+  ...serviceAndServiceCategoryTypeDefs,
+  ...lbmAndServiceTypeDefs,
   ...lbmAndLbmTypeTypeDefs,
   // ...lbmAndPersonTypeDefs,
   // ...lbmAndKeyEvent,