+++ src/frontend/api/mutations/organization.js
@@ -407,12 +407,35 @@ export const CONNECT_OBM_SERVICE_AND_OBM_SERVICE_CATEGORY = gql`
   }
 `
 
+export const CONNECT_LBM_SERVICE_AND_LBM_SERVICE_CATEGORY = gql`
+  mutation ConnectLbmServiceAndLbmServiceCategory(
+    $input: ConnectLbmServiceAndLbmServiceCategoryInput!
+  ) {
+    connectLbmServiceAndLbmServiceCategory(input: $input) {
+      _id
+      lbmServiceId
+      lbmServiceCategoryId
+    }
+  }
+`
+
 export const CONNECT_OBM_AND_OBM_SERVICE = gql`
   mutation ConnectObmAndObmService($input: [ConnectObmAndObmServiceInput!]!) {
     connectObmAndObmService(input: $input) {
       _id
       obmId
       obmServiceId
+      rating
+    }
+  }
+`
+
+export const CONNECT_LBM_AND_LBM_SERVICE = gql`
+  mutation ConnectLbmAndLbmService($input: [ConnectLbmAndLbmServiceInput!]!) {
+    connectLbmAndLbmService(input: $input) {
+      _id
+      lbmId
+      lbmServiceId
       rating
     }
   }